rename conflict model
