--- packages/app/src/server/models/conflicts.ts
@@ -1,50 +1,0 @@
-import { truncate } from 'fs/promises';
-import { Types, Schema } from 'mongoose';
-import uniqueValidator from 'mongoose-unique-validator';
-
-interface IRevisionOnConflict {
-  id: string,
-  body: string,
-  userName: string,
-  userImgPath: string,
-  createdAt: Date;
-}
-
-type IConflict = {
-  path: string;
-  revisions: {
-    request: IRevisionOnConflict,
-    origin: IRevisionOnConflict,
-    latest: IRevisionOnConflict,
-  };
-};
-
-
-const schema = new Schema<IConflict>({
-  path: { type: String, required: true, uniquer: true },
-  revisions: {
-    request: {
-      id: { type: String },
-      body: { type: String, required: true },
-      userName: { type: String, required: true },
-      userImgPath: { type: String, required: true },
-      createdAt: { type: Date, required: true },
-    },
-    origin: {
-      id: { type: String },
-      body: { type: String, required: true },
-      userName: { type: String, required: true },
-      userImgPath: { type: String, required: true },
-      createdAt: { type: Date, required: truncate },
-    },
-    latest: {
-      id: { type: String },
-      body: { type: String, required: true },
-      userName: { type: String, required: true },
-      userImgPath: { type: String, required: true },
-      createdAt: { type: Date, required: true },
-    },
-  },
-});
-
-schema.plugin(uniqueValidator);
